fixed type def erro

+++ tgdl_solver.py
@@ -25,14 +25,14 @@ grid_length = 100.
 # Define number of grid points (int)
 n_points = 500
 # Define Point spacing
-dx = grid_length / n_points
+dx = np.float32(grid_length / n_points)
 # Set time step
-delta_time = 0.01 * dx ** 2.
+delta_time = np.float32(0.01 * dx ** 2.)
 
 neighbours = np.zeros((2, n_points, 4), dtype=np.int64)
 tf.get_all_neighbours(neighbours, n_points)
 
-sample_times = np.linspace(0, 50, 500)
+sample_times = np.linspace(0, 50, 500, dtype=np.float32)
 
 
 # Start the clock to estimate the total time taken for this lattice size
